moved into own class
