+++ src/main/java/at/hannibal2/skyhanni/data/GardenCropMilestones.kt
@@ -1,33 +1,21 @@
 package at.hannibal2.skyhanni.data
 
-import at.hannibal2.skyhanni.SkyHanniMod
-import at.hannibal2.skyhanni.config.ConfigManager
 import at.hannibal2.skyhanni.events.CropMilestoneUpdateEvent
 import at.hannibal2.skyhanni.events.InventoryFullyOpenedEvent
 import at.hannibal2.skyhanni.events.RepositoryReloadEvent
 import at.hannibal2.skyhanni.features.garden.CropType
 import at.hannibal2.skyhanni.features.garden.GardenAPI
 import at.hannibal2.skyhanni.utils.ItemUtils.getLore
-import at.hannibal2.skyhanni.utils.ItemUtils.name
-import at.hannibal2.skyhanni.utils.LorenzUtils
-import at.hannibal2.skyhanni.utils.LorenzUtils.editCopy
-import at.hannibal2.skyhanni.utils.LorenzUtils.nextAfter
-import at.hannibal2.skyhanni.utils.NumberUtil.addSeparators
 import at.hannibal2.skyhanni.utils.NumberUtil.formatNumber
-import at.hannibal2.skyhanni.utils.NumberUtil.romanToDecimalIfNeeded
-import at.hannibal2.skyhanni.utils.OSUtils
 import at.hannibal2.skyhanni.utils.StringUtils.matchMatcher
-import at.hannibal2.skyhanni.utils.StringUtils.matches
-import at.hannibal2.skyhanni.utils.StringUtils.removeColor
 import at.hannibal2.skyhanni.utils.jsonobjects.GardenJson
-import kotlinx.coroutines.launch
 import net.minecraft.item.ItemStack
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
 
 object GardenCropMilestones {
     // TODO USE SH-REPO
     private val cropPattern = "§7Harvest §f(?<name>.*) §7on .*".toPattern()
-    private val totalPattern = "§7Total: §a(?<name>.*)".toPattern()
+    val totalPattern = "§7Total: §a(?<name>.*)".toPattern()
 
     fun getCropTypeByLore(itemStack: ItemStack): CropType? {
         for (line in itemStack.getLore()) {
@@ -53,76 +41,10 @@ object GardenCropMilestones {
             }
         }
         CropMilestoneUpdateEvent().postAndCatch()
-        if (SkyHanniMod.feature.garden.copyMilestoneData) {
-            fixForWrongData(event.inventoryItems)
-        }
+        GardenCropMilestonesFix.openInventory(event.inventoryItems)
     }
 
-    private fun fixForWrongData(inventoryItems: Map<Int, ItemStack>) {
-        val data = mutableListOf<String>()
-        for ((_, stack) in inventoryItems) {
-            val crop = getCropTypeByLore(stack) ?: continue
-            checkForWrongData(stack, crop, data)
-
-            CropMilestoneUpdateEvent().postAndCatch()
-        }
-
-        if (data.isNotEmpty()) {
-            LorenzUtils.chat(
-                "Found §c${data.size} §ewrong crop milestone steps in the menu! " +
-                    "Correct data got put into clipboard. " +
-                    "Please share it on SkyHanni Discord."
-            )
-            OSUtils.copyToClipboard("```${data.joinToString("\n")}```")
-        }
-    }
-
-    private fun checkForWrongData(
-        stack: ItemStack,
-        crop: CropType,
-        wrongData: MutableList<String>
-    ) {
-        val pattern = ".*§e(?<having>.*)§6/§e(?<max>.*)".toPattern()
-        val name = stack.name ?: return
-        val rawNumber = name.removeColor().replace(crop.cropName, "").trim()
-        val realTier = if (rawNumber == "") 0 else rawNumber.romanToDecimalIfNeeded()
-
-        val lore = stack.getLore()
-        val next = lore.nextAfter({ totalPattern.matches(it) }, 3) ?: return
-        val total = lore.nextAfter({ totalPattern.matches(it) }, 6) ?: return
-
-        debug(" ")
-        debug("crop: $crop")
-        debug("realTier: $realTier")
-
-        val guessNextMax = getCropsForTier(realTier + 1, crop) - getCropsForTier(realTier, crop)
-        debug("guessNextMax: ${guessNextMax.addSeparators()}")
-        val nextMax = pattern.matchMatcher(next) {
-            group("max").formatNumber()
-        } ?: return
-        debug("nextMax real: ${nextMax.addSeparators()}")
-        if (nextMax != guessNextMax) {
-            debug("wrong, add to list")
-            wrongData.add("$crop:$realTier:${nextMax.addSeparators()}")
-        }
-
-        val guessTotalMax = getCropsForTier(46, crop)
-        //             println("guessTotalMax: ${guessTotalMax.addSeparators()}")
-        val totalMax = pattern.matchMatcher(total) {
-            group("max").formatNumber()
-        } ?: return
-        //             println("totalMax real: ${totalMax.addSeparators()}")
-        val totalOffBy = guessTotalMax - totalMax
-        debug("totalOffBy: $totalOffBy")
-    }
-
-    fun debug(message: String) {
-        if (SkyHanniMod.feature.dev.debug.enabled) {
-//             println(message)
-        }
-    }
-
-    private var cropMilestoneData: Map<CropType, List<Int>> = emptyMap()
+    var cropMilestoneData: Map<CropType, List<Int>> = emptyMap()
 
     val cropCounter: MutableMap<CropType, Long>? get() = GardenAPI.storage?.cropCounter
 
@@ -182,54 +104,5 @@ object GardenCropMilestones {
     @SubscribeEvent
     fun onRepoReload(event: RepositoryReloadEvent) {
         cropMilestoneData = event.getConstant<GardenJson>("Garden").crop_milestones
-    }
-
-    fun readDataFromClipboard() {
-        SkyHanniMod.coroutineScope.launch {
-            OSUtils.readFromClipboard()?.let {
-                handleInput(it)
-            }
-        }
-    }
-
-    private var totalFixedValues = 0
-
-    fun handleInput(input: String) {
-        println(" ")
-        var fixed = 0
-        var alreadyCorrect = 0
-        for (line in input.lines()) {
-            val split = line.replace("```", "").replace(".", ",").split(":")
-            if (split.size != 3) continue
-            val (rawCrop, tier, amount) = split
-            val crop = LorenzUtils.enumValueOf<CropType>(rawCrop)
-
-            if (tryFix(crop, tier.toInt(), amount.formatNumber().toInt())) {
-                fixed++
-            } else {
-                alreadyCorrect++
-            }
-        }
-        totalFixedValues += fixed
-        LorenzUtils.chat("Fixed: $fixed/$alreadyCorrect, total fixes: $totalFixedValues")
-        val s = ConfigManager.gson.toJsonTree(cropMilestoneData).toString()
-        OSUtils.copyToClipboard("\"crop_milestones\":$s,")
-    }
-
-    private fun tryFix(crop: CropType, tier: Int, amount: Int): Boolean {
-        val guessNextMax = getCropsForTier(tier + 1, crop) - getCropsForTier(tier, crop)
-        if (guessNextMax.toInt() == amount) {
-            return false
-        }
-        cropMilestoneData = cropMilestoneData.editCopy {
-            fix(crop, this, tier, amount)
-        }
-        return true
-    }
-
-    private fun fix(crop: CropType, map: MutableMap<CropType, List<Int>>, tier: Int, amount: Int) {
-        map[crop] = map[crop]!!.editCopy {
-            this[tier] = amount
-        }
     }
 }
